[costflow] moving underlying code to correct subdirectories
