+++ include/chinese/pgr_chinesePostman.hpp
@@ -35,7 +35,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <limits>
 #include <stack>
 
-#include "costFlow/pgr_minCostMaxFlow.hpp"
+#include "max_flow/pgr_minCostMaxFlow.hpp"
 #include "c_types/general_path_element_t.h"
 #include "c_types/pgr_edge_t.h"
 #include "c_types/pgr_flow_t.h"
