--- src/costFlow/CMakeLists.txt
@@ -1,5 +1,0 @@
-ADD_LIBRARY(costFlow OBJECT
-    minCostMaxFlow.c
-    minCostMaxFlow_driver.cpp
-
-    pgr_minCostMaxFlow.cpp)
